Move PlaylistData to a dedicated file
