+++ yledl/data_extractors.py
@@ -1,9 +1,8 @@
-from typing import Dict, List, Optional
+from typing import List, Optional
 from datetime import datetime
 
 from dataclasses import dataclass, field
 from .areena_playlist_parser import AreenaPlaylistParser
-from .http import update_url_query
 from .io import OutputFileNameGenerator
 from .streamflavor import StreamFlavor
 from .subtitles import Subtitle
@@ -126,22 +125,6 @@ class AreenaApiProgramInfo:
     pending: bool
     expired: bool
 
-
-@dataclass(frozen=True)
-class PlaylistData:
-    # The base URL from which to download a playlist
-    base_url: str
-    # List of query parameters. Each item is a dictionary of query
-    # parameters for one season. If empty, a playlist is downloaded
-    # from the plain base_url.
-    season_parameters: List[Dict]
-
-    def season_playlist_urls(self):
-        if self.season_parameters:
-            for season_query in self.season_parameters:
-                yield update_url_query(self.base_url, season_query)
-        else:
-            yield self.base_url
 
 
 @dataclass(frozen=True)
